bar bars more now, but it still doesn't bar healthily

+++ Subject Escape/Assets/Scripts/PlayerHealth.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class PlayerHealth : MonoBehaviour
 {
     public GameObject Player;
-    public float Health = 100;
+    public int Health = 100;
     // Start is called before the first frame update
     void Start()
     {
